+++ lib/carelink/cli/csv_loader.js
@@ -5,7 +5,7 @@ var fs = require('fs');
 
 var driverManager = require('../../driverManager');
 var pwdSimulator = require('../carelinkSimulator.js');
-var carelinkDriver = require('../carelinkDriver');
+var carelinkDriver = require('../../drivers/carelinkDriver');
 var api = require('../../core/api.js');
 var config = require('../../../.config.js');
 var pkg = require('./../../../package.json');
