+++ lib/core/carelink.js
@@ -19,7 +19,7 @@ var _ = require('lodash');
 
 var driverManager = require('../driverManager');
 var pwdSimulator = require('../carelink/carelinkSimulator.js');
-var carelinkDriver = require('../carelink/carelinkDriver');
+var carelinkDriver = require('../drivers/carelinkDriver');
 
 var carelink = {
   log: require('../bows')('Carelink')
