+++ lib/drivers/carelinkDriver.js
@@ -22,9 +22,9 @@ var util = require('util');
 var csv = require('babyparse');
 var sundial = require('sundial');
 
-var common = require('./common');
+var common = require('../carelink/common');
 var debug = require('../bows')('CarelinkDriver');
-var removeOverlaps = require('./removeOverlapping');
+var removeOverlaps = require('../carelink/removeOverlapping');
 var CARELINK_TS_FORMAT = 'MM/DD/YY HH:mm:ss';
 
 function convertRawValues(e) {
@@ -53,12 +53,12 @@ function initializeProcessors(timezone, model) {
   // are visited determines the order that events will be delivered to the simulator.  Keep this in mind when
   // re-ordering.
   return [
-    require('./settings.js')(timezone),
-    require('./smbg.js')(timezone),
-    require('./cbg.js')(timezone),
-    require('./basal.js')(timezone),
-    require('./suspend.js')(timezone),
-    require('./bolusNWizard')(timezone, model)
+    require('../carelink/settings.js')(timezone),
+    require('../carelink/smbg.js')(timezone),
+    require('../carelink/cbg.js')(timezone),
+    require('../carelink/basal.js')(timezone),
+    require('../carelink/suspend.js')(timezone),
+    require('../carelink/bolusNWizard')(timezone, model)
   ];
 }
