+++ lib/insulet/cli/ibf_loader.js
@@ -4,7 +4,7 @@ var program = require('commander');
 var fs = require('fs');
 
 var driverManager = require('../../driverManager');
-var insuletDriver = require('../insuletDriver.js');
+var insuletDriver = require('../../drivers/insuletDriver.js');
 var api = require('../../core/api.js');
 var config = require('../../../.config.js');
 var pkg = require('./../../../package.json');
